+++ java_test/java/jtest/Main.java
@@ -10,6 +10,7 @@ import tech.v3.datatype.NDBuffer;
 import static tech.v3.Clj.*;
 import static tech.v3.DType.*;
 import static tech.v3.Tensor.*;
+import tech.v3.libs.Nippy;
 import tech.v3.datatype.VecMath;
 import tech.v3.datatype.Stats;
 import tech.v3.DType; //explicit access to clone method
@@ -52,18 +53,18 @@ public class Main
     ArrayList<Object> data = new ArrayList<Object>();
     data.addAll(buf);
     data.addAll(dbuf);
-    System.out.println("Readers: " + data.toString());
-    System.out.println("Last Elem: " + call(buf, -1).toString()
+    println("Readers: " + data.toString());
+    println("Last Elem: " + call(buf, -1).toString()
 		       + " " + call(dbuf,-1).toString()
 		       + " " + apply(buf, -1).toString());
     //Readers: [0, 1, 2, 0.0, 1.0, 2.0]
     //Last Elem: 2 2.0 2
 
     //'Vectorized' math is provided
-    System.out.println(vec(VecMath.add(dbuf,dbuf)).toString() + " : "
+    println(vec(VecMath.add(dbuf,dbuf)).toString() + " : "
 		       + VecMath.magnitudeSquared(dbuf));
 
-    System.out.println(vec(VecMath.where(range(10)
+    println(vec(VecMath.where(range(10)
 					 , new IFnDef() {
 					     public Object invoke(Object val) {
 					       return RT.longCast(val) > 5;
@@ -76,7 +77,7 @@ public class Main
     //into a tensor.  Arrays-of-arrays need to go through the makeTensor route.
     int[] ddata = new int[] {0,1,2,3,4,5,6,7,8,9};
 
-    System.out.println(call(reshape, ddata, vector(3,3)).toString());
+    println(call(reshape, ddata, vector(3,3)).toString());
     //#tech.v3.tensor<int64>[3 3]
     //[[0 1 2]
     // [3 4 5]
@@ -86,7 +87,7 @@ public class Main
     //Clojure libraries often use maps with keyword keys as optional arguments to functions.
     //opts allows you to construct one of these maps without needing to type the
     //keyword function out repeatedly.
-    System.out.println(opts("one", 1, "two", 2, "three", 3).toString());
+    println(opts("one", 1, "two", 2, "three", 3).toString());
     //{:one 1, :two 2, :three 3}
 
     //When allocating native memory, users can control how the memory is reclaimed.
@@ -99,12 +100,12 @@ public class Main
     try (AutoCloseable ac = stackResourceContext()) {
       Object nativeBuf = makeContainer(nativeHeap, int8, opts("log-level", keyword("info")),
 				       range(10));
-      System.out.println(nativeBuf.toString());
+      println(nativeBuf.toString());
     } catch (Exception e) {
-      System.out.println("Error!!" + e.toString());
+      println("Error!!" + e.toString());
       e.printStackTrace(System.out);
     }
-    System.out.println("After stack pop - nativemem should be released");
+    println("After stack pop - nativemem should be released");
     //Jan 17, 2022 9:05:39 AM clojure.tools.logging$eval3217$fn__3220 invoke
     //INFO: Malloc - 0x00007F37B91C7D90 - 0000000000000010 bytes
     //#native-buffer@0x00007F37B91C7D90<int8>[10]
@@ -122,30 +123,30 @@ public class Main
     Object srcbuf = makeContainer(float32, range(10));
     Object dstbuf = makeContainer(float32, 10);
     copy(srcbuf, dstbuf);
-    System.out.println(dstbuf.toString());
+    println(dstbuf.toString());
     //#array-buffer<float32>[10]
     //[0.000, 1.000, 2.000, 3.000, 4.000, 5.000, 6.000, 7.000, 8.000, 9.000]
 
     //setConstant is a high performance primitive when possible.
     setConstant(dstbuf, 0);
-    System.out.println(dstbuf.toString());
+    println(dstbuf.toString());
     //#array-buffer<float32>[10]
     //[0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000]
 
     //reshape allows you to reshape a 1D buffer or an ND buffer interpreted row-major as
     //1D into an ND tensor.
-    System.out.println(reshape(range(10), vector(3,3)).toString());
+    println(reshape(range(10), vector(3,3)).toString());
     //#tech.v3.tensor<int64>[3 3]
     //[[0 1 2]
     // [3 4 5]
     // [6 7 8]]
 
-    System.out.println(reshape(toDoubleArray(range(9)), vector(3,3)).toString());
+    println(reshape(toDoubleArray(range(9)), vector(3,3)).toString());
     //#tech.v3.tensor<float64>[3 3]
     //[[0.000 1.000 2.000]
     // [3.000 4.000 5.000]
     // [6.000 7.000 8.000]]
-    System.out.println(computeTensor(float64, vector(3,3),
+    println(computeTensor(float64, vector(3,3),
 				     new IFnDef() {
 				       public Object invoke(Object yidx, Object xidx) {
 					 return (RT.longCast(yidx) * 3) + RT.longCast(xidx);
@@ -157,28 +158,28 @@ public class Main
     // [6.000 7.000 8.000]]
     try (AutoCloseable ac = stackResourceContext()) {
       Map bufDesc = ensureNDBufferDescriptor(reshape(toDoubleArray(range(9)), vector(3,3)));
-      System.out.println(bufDesc.toString());
+      println(bufDesc.toString());
       //{:ptr 139877299155184, :datatype :tensor, :elemwise-datatype :float64, :endianness :little-endian, :shape [3 3], :strides [24 8], :native-buffer #native-buffer@0x00007F37B8BA70F0<float64>[9]
       //[0.000, 1.000, 2.000, 3.000, 4.000, 5.000, 6.000, 7.000, 8.000]}
-      System.out.println(NDBufferDescriptorToTensor(bufDesc).toString());
+      println(NDBufferDescriptorToTensor(bufDesc).toString());
       //#tech.v3.tensor<float64>[3 3]
       //[[0.000 1.000 2.000]
       // [3.000 4.000 5.000]
       // [6.000 7.000 8.000]]
     } catch(Exception e){
-      System.out.println(e.toString());
+      println(e.toString());
       e.printStackTrace(System.out);
     }
 
     //Ensure nio buffers are supported
     FloatBuffer fbuf = FloatBuffer.wrap(toFloatArray(range (9)));
-    System.out.println(reshape(fbuf, vector(3,3)).toString());
+    println(reshape(fbuf, vector(3,3)).toString());
     //#tech.v3.tensor<float32>[3 3]
     // [[0.000 1.000 2.000]
     //  [3.000 4.000 5.000]
     //  [6.000 7.000 8.000]]
     fbuf = FloatBuffer.allocate(9);
-    System.out.println(reshape(fbuf, vector(3,3)).toString());
+    println(reshape(fbuf, vector(3,3)).toString());
     //#tech.v3.tensor<float32>[3 3]
     //[[0.000 0.000 0.000]
     // [0.000 0.000 0.000]
@@ -191,7 +192,7 @@ public class Main
     Buffer wrappedData = toBuffer(testData);
     ArrayBufferData origData = asArrayBuffer(wrappedData);
 
-    System.out.println(String.valueOf(System.identityHashCode(testData))
+    println(String.valueOf(System.identityHashCode(testData))
 		       + " " + String.valueOf(System.identityHashCode(origData.arrayData)));
 
 
@@ -199,9 +200,9 @@ public class Main
     //incrementing by 1.
     NDBuffer tensData = reshape(toDoubleArray(range (27)), vector(3,3,3));
     NDBuffer smallerTensor = select(tensData, range(1,3));
-    System.out.println(smallerTensor.toString());
+    println(smallerTensor.toString());
     //Because we use a range the above condition has to be met.
-    System.out.println("Array buffer access? " +
+    println("Array buffer access? " +
 		       (String.valueOf(asArrayBuffer(smallerTensor) != null)));
     // #tech.v3.tensor<float64>[2 3 3]
     // [[[9.000 10.00 11.00]
@@ -215,9 +216,9 @@ public class Main
 
     //If we invert the selection thus reversing the outermost dimension
     smallerTensor = select(tensData, range(2,0,-1));
-    System.out.println(smallerTensor.toString());
+    println(smallerTensor.toString());
     //We can no longer get a buffer from this tensor.
-    System.out.println("Array buffer access? " +
+    println("Array buffer access? " +
 		       (String.valueOf(asArrayBuffer(smallerTensor) != null)));
     // #tech.v3.tensor<float64>[2 3 3]
     // [[[18.00 19.00 20.00]
@@ -229,13 +230,13 @@ public class Main
     // Array buffer access? false
 
     //makeTensor copies data into the tensor.
-    System.out.println(makeTensor(range(9), vector(3,3), float32).toString());
+    println(makeTensor(range(9), vector(3,3), float32).toString());
     //#tech.v3.tensor<float32>[3 3]
     // [[0.000 1.000 2.000]
     //  [3.000 4.000 5.000]
     //  [6.000 7.000 8.000]]
     //We can also make tensor's based on native-heap memory.
-    System.out.println(makeTensor(range(9), vector(3,3), float32, nativeHeap).toString());
+    println(makeTensor(range(9), vector(3,3), float32, nativeHeap).toString());
     //#tech.v3.tensor<float32>[3 3]
     // [[0.000 1.000 2.000]
     //  [3.000 4.000 5.000]
@@ -244,9 +245,27 @@ public class Main
     //Native memory tensors have an in-place conversion to a native buffer.
     NDBuffer ntens = makeTensor(range(9), vector(3,3), float32, nativeHeap);
     NativeBufferData ndata = asNativeBuffer(ntens);
-    System.out.println(ndata.address);
+    println(ndata.address);
     //140397742494704
 
+
+    //We can also serialize arbitrary Clolure datastructures to an
+    //efficient and compressed format called 'nippy'.
+    //Below could be a single tensor, a map of key to tensor
+    //OR, with the dataset bindings loaded, a heterogeneous map where
+    //some of the values are tensors and some of the values are datasets.
+    byte[] frozenTens = Nippy.freeze(tensData);
+    println(Nippy.thaw(frozenTens));
+    //#tech.v3.tensor<float64>[3 3 3]
+    //[[[0.000 1.000 2.000]
+    //  [3.000 4.000 5.000]
+    //  [6.000 7.000 8.000]]
+    // [[9.000 10.00 11.00]
+    //  [12.00 13.00 14.00]
+    //  [15.00 16.00 17.00]]
+    // [[18.00 19.00 20.00]
+    //  [21.00 22.00 23.00]
+    //  [24.00 25.00 26.00]]]
 
 
     //Neanderthal is a high speed linear algebra system with bindings to MKL, openCL, and
@@ -261,27 +280,40 @@ public class Main
     require("uncomplicate.neanderthal.core");
     IFn denseConstructor = requiringResolve("uncomplicate.neanderthal.native", "dge");
     Object denseMatrix = call(denseConstructor, 3, 3, range(9));
-    System.out.println(denseMatrix.toString());
-    // #RealGEMatrix[double, mxn:3x3, layout:column, offset:0]
+    println(denseMatrix);
+    //#RealGEMatrix[double, mxn:3x3, layout:column, offset:0]
+    //     ▥       ↓       ↓       ↓       ┓
+    //     →       0.00    3.00    6.00
+    //     →       1.00    4.00    7.00
+    //     →       2.00    5.00    8.00
+    //     ┗                               ┛
 
     //Neaderthal support for dtype-next requires you to require a namespace.  Then
     //protocols are auto-loaded and neaderthal dense matrixes will have a as-tensor pathway.
-    require("tech.v3.libs.neanderthal");
+    //Tensor.enableNeanderthal
+    enableNeanderthal();
     //in-place conversion into dtype-next land is then supported.
     NDBuffer ndBuf = asTensor(denseMatrix);
     //Neaderthal defaults to column-major storage.
-    System.out.println(ndBuf.toString());
+    println(ndBuf.toString());
     //#tech.v3.tensor<float64>[3 3]
     // [[0.000 3.000 6.000]
     //  [1.000 4.000 7.000]
     //  [2.000 5.000 8.000]]
+    println(tensorToNeanderthalMatrix(ndBuf));
+    //#RealGEMatrix[double, mxn:3x3, layout:column, offset:0]
+    //     ▥       ↓       ↓       ↓       ┓
+    //     →       0.00    3.00    6.00
+    //     →       1.00    4.00    7.00
+    //     →       2.00    5.00    8.00
+    //     ┗                               ┛
 
     //We can create row-major matrix by passing in an option map.
     Object denseMatrixRowMajor = call(denseConstructor, 3, 3, range(9),
 				      opts("layout", kw("row")));
 
     ndBuf = asTensor(denseMatrixRowMajor);
-    System.out.println(ndBuf.toString());
+    println(ndBuf);
     //#tech.v3.tensor<float64>[3 3]
     //[[0.000 1.000 2.000]
     // [3.000 4.000 5.000]
@@ -305,7 +337,7 @@ public class Main
     //regardless you can always print a dataset safely without bombing your print system.
     //Datasets print in markdown table format so you can paste them directly into a markdown
     //document if you want a table.
-    System.out.println(ds.toString());
+    println(ds.toString());
     //https://github.com/techascent/tech.ml.dataset/raw/master/test/data/stocks.csv [560 3]:
 
     //| symbol |       date | price |
@@ -325,13 +357,13 @@ public class Main
     //Columns aren't themselves readers specifically but they do implement IFn expecting
     //indexes and they also print nicely.
     Object prices = dsMap.get("price");
-    System.out.println(prices);
+    println(prices);
     //#tech.v3.dataset.column<float64>[560]
     //price
     //[39.81, 36.35, 43.22, 28.37, 25.45, 32.54, 28.40, 28.40, 24.53, 28.02, 23.34, 17.65, 24.84, 24.00, 22.25, 27.56, 28.14, 29.70, 26.93, 23.21...]
 
     //Print the last price
-    System.out.println("Last price: " + call(prices, -1).toString());
+    println("Last price: " + call(prices, -1).toString());
     //Last price: 223.02
 
     //Datasets support a few java.time objects - Instants, LocalDates, and LocalTimes.  These
@@ -340,12 +372,12 @@ public class Main
     //epoch-days and local-times are stores as 64bit microseconds past midnight.
     //These specific datatypes are prefixed with :packed-.
     Object dates = dsMap.get("date");
-    System.out.println("Packed date datatype: " + elemwiseDatatype(dates).toString());
+    println("Packed date datatype: " + elemwiseDatatype(dates).toString());
     //Packed date datatype: :packed-local-date
 
     //Reading from this column's readObject member or its IFn interface will get you back
     //a LocalDate
-    System.out.println(call(dates, 0).toString());
+    println(call(dates, 0).toString());
 
 
     //Datatype contains a namespace with math functions that are designed to work
@@ -356,7 +388,7 @@ public class Main
     Object sqFn = requiringResolve("tech.v3.datatype.functional", "sq");
     Object headFn = requiringResolve("tech.v3.dataset", "head");
     Object newDs = assoc(ds, "price^2", call(sqFn, prices));
-    System.out.println(call(headFn, newDs).toString());
+    println(call(headFn, newDs).toString());
     //https://github.com/techascent/tech.ml.dataset/raw/master/test/data/stocks.csv [5 4]:
 
     //| symbol |       date | price |   price^2 |
@@ -378,10 +410,10 @@ public class Main
     //The easiest pathway IMHO is to use the IFn or Buffer readObject interface as then
     //you get negative from-the-end indexing.
     Object rows = (Object)call(rowsFn, newDs);
-    System.out.println(call(rows, 0).toString());
+    println(call(rows, 0).toString());
     // {"date" #object[java.time.LocalDate 0xe360e01 "2000-01-01"], "symbol" "MSFT", "price^2" 1584.8361000000002, "price" 39.81}
 
-    System.out.println(call(rows,-1).toString());
+    println(call(rows,-1).toString());
     // {"date" #object[java.time.LocalDate 0x210588 "2010-03-01"], "symbol" "AAPL", "price^2" 49737.9204, "price" 223.02}
 
     //Saving/loading to arrow is done through a specific namespace.
@@ -400,7 +432,7 @@ public class Main
       //If open-type isn't provided it uses the much more robust input-stream
       //pathway.
       Object mmapds = call(readFn, "test.arrow", opts("open-type", kw("mmap")));
-      System.out.println(call(headFn, mmapds).toString());
+      println(call(headFn, mmapds).toString());
       //test.arrow [5 4]:
       //| symbol |       date | price |   price^2 |
       //|--------|------------|------:|----------:|
@@ -417,7 +449,7 @@ public class Main
       //as efficiently as possible into jvmHeap memory.
       //Cloning also serves to realize any outstanding lazy operations.
       Map cloneds = (Map)DType.clone(mmapds);
-      System.out.println(call(headFn, cloneds).toString());
+      println(call(headFn, cloneds).toString());
       //test.arrow [5 4]:
       //| symbol |       date | price |   price^2 |
       //|--------|------------|------:|----------:|
@@ -427,7 +459,7 @@ public class Main
       //|   MSFT | 2000-04-01 | 28.37 |  804.8569 |
       //|   MSFT | 2000-05-01 | 25.45 |  647.7025 |
     } catch(Exception e){
-      System.out.println(e.toString());
+      println(e.toString());
       e.printStackTrace(System.out);
     }
 
@@ -436,7 +468,7 @@ public class Main
     HashMap dsInput = new HashMap();
     dsInput.put("doubles", toDoubleArray(range(10)));
     dsInput.put("shorts", toShortArray(range(9, -1, -1)));
-    System.out.println(call(datasetConstructor, dsInput).toString());
+    println(call(datasetConstructor, dsInput).toString());
     //_unnamed [10 2]:
     //| doubles | shorts |
     //|--------:|-------:|
@@ -497,7 +529,7 @@ public class Main
 				   return sum;
 				 }
 			       });
-    System.out.println("Reduction result: " + String.valueOf(result));
+    println("Reduction result:", result);
     //Reduction result: 4.999995E11
 
     //Now that was a long form super high efficiency summation.  Now we trade a bit of
@@ -505,8 +537,7 @@ public class Main
     //NaN values and second we use Kahan's compensated summation algorithm to give the
     //accumulator more effective bits than 64.
 
-    System.out.println("Built in reduction result: " +
-		       String.valueOf(Stats.sum(doubles)));
+    println("Built in reduction result:", Stats.sum(doubles));
     //Built in reduction result: 4.999995E11
 
     //Neanderthal boots up Clojure's agent pool which means that when it comes time to