+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+## 9.004
+ * Expose tensor to neanderthal and nippy freeze/thaw of tensors.
+ 
 ## 9.002
- * expose map,filter from clj interface.
+ * Expose map,filter from clj interface.
  
 ## 9.001
  * Added tech.v3.datatype.Pred package to allow fast (in terms of typing) creation of typed